feat: added a new feature of keyboard type in input component
