--- src/components/text-input/index.ts
@@ -1,2 +1,0 @@
-export { default } from './text-input';
-export { KeyboardTypes } from '../../types/text-input';
