+++ src/types/index.ts
@@ -5,3 +5,4 @@ export * from './auth';
 export * from './common';
 export * from './task';
 export * from './avatar';
+export * from './input';
